test: CarTest, RacingControllerTest 분리
